Refactor carbon-core module for code clean up (#720)
* refactor core module

* fix style
