+++ core/src/main/java/org/carbondata/scan/executor/QueryExecutor.java
@@ -18,7 +18,7 @@
  */
 package org.carbondata.scan.executor;
 
-import org.carbondata.core.iterator.CarbonIterator;
+import org.carbondata.common.CarbonIterator;
 import org.carbondata.scan.executor.exception.QueryExecutionException;
 import org.carbondata.scan.model.QueryModel;
