+++ processing/src/main/java/org/carbondata/lcm/fileoperations/AtomicFileOperations.java
@@ -17,9 +17,17 @@
  * under the License.
  */
 
-package org.carbondata.core.datastorage.store.fileperations;
+package org.carbondata.lcm.fileoperations;
 
-public enum FileWriteOperation {
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
 
-  APPEND, OVERWRITE
+public interface AtomicFileOperations {
+
+  DataInputStream openForRead() throws IOException;
+
+  void close() throws IOException;
+
+  DataOutputStream openForWrite(FileWriteOperation operation) throws IOException;
 }
