+++ processing/src/main/java/org/carbondata/lcm/locks/AbstractCarbonLock.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.carbondata.core.locks;
+package org.carbondata.lcm.locks;
 
 import org.carbondata.core.constants.CarbonCommonConstants;
 import org.carbondata.core.util.CarbonProperties;
