+++ processing/src/main/java/org/carbondata/lcm/locks/LocalFileLock.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.carbondata.core.locks;
+package org.carbondata.lcm.locks;
 
 import java.io.File;
 import java.io.FileOutputStream;
