+++ processing/src/main/java/org/carbondata/lcm/locks/LockUsage.java
@@ -16,23 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package org.carbondata.core.iterator;
-
-import java.util.Iterator;
+package org.carbondata.lcm.locks;
 
 /**
- * CarbonIterator adds default implement for remove. This is required for Java 7.
- * @param <E>
+ * This enum is used to define the usecase of the lock.
+ * Each enum value is one specific lock case.
  */
-public abstract class CarbonIterator<E> implements Iterator<E> {
-
-  @Override public abstract boolean hasNext();
-
-  @Override public abstract E next();
-
-  @Override public void remove() {
-    throw new UnsupportedOperationException("remove");
-  }
+public enum LockUsage {
+  METADATA_LOCK,
+  COMPACTION_LOCK;
 
 }
