+++ processing/src/main/java/org/carbondata/lcm/locks/ZookeeperInit.java
@@ -1,4 +1,4 @@
-package org.carbondata.core.locks;
+package org.carbondata.lcm.locks;
 
 import java.io.IOException;
