+++ processing/src/main/java/org/carbondata/processing/mdkeygen/file/IFileManagerComposite.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.carbondata.core.file.manager.composite;
+package org.carbondata.processing.mdkeygen.file;
 
 public interface IFileManagerComposite {
   /**
