+++ processing/src/test/java/org/carbondata/lcm/locks/ZooKeeperLockingTest.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package org.carbondata.core.locks;
+package org.carbondata.lcm.locks;
 
 import java.io.File;
 import java.io.IOException;
